requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==3.2
django-cors-headers==3.14.0
djongo==1.3.6
dnspython==2.3.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-language==2.9.1
google-cloud-speech==2.19.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphene==3.2.2
graphene-django==3.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
promise==2.3
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pymongo==3.12.3
python-decouple==3.8
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
